pyproject.toml
[tool.poetry]
name = "fastapi_dev"
version = "0.1.0"
description = ""
authors = ["alexander <alfacat187@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standart"], version = "^0.112.2"}
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.32"
pydantic = {extras = ["email"], version = "^2.8.2"}
alembic = "^1.13.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
